file_processing_website/backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库和缓存
redis[asyncio]==5.0.1

# 异步任务
celery==5.3.4

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0

# 文件处理
aiofiles==23.2.1
python-magic==0.4.27

# 日志
loguru==0.7.2

# 工具库
python-dotenv==1.0.0
httpx==0.25.2

# CORS支持 (FastAPI内置)

# WebSocket支持
websockets==12.0

# 时间处理
python-dateutil==2.8.2

# JSON处理
orjson==3.9.10 